requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
altair==4.2.0
anyio==3.6.2
appdirs==1.4.4
async-timeout==4.0.2
asyncer==0.0.2
attrs==22.2.0
blinker==1.5
cachetools==5.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
commonmark==0.9.1
decorator==5.1.1
entrypoints==0.4
fastapi==0.87.0
filetype==1.2.0
flatbuffers==23.1.4
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.30
h11==0.14.0
humanfriendly==10.0
idna==3.4
ImageHash==4.3.1
imageio==2.24.0
importlib-metadata==6.0.0
Jinja2==3.1.2
jsonschema==4.17.3
llvmlite==0.39.1
MarkupSafe==2.1.2
mpmath==1.2.1
multidict==6.0.4
networkx==3.0
numba==0.56.4
numpy==1.23.5
onnxruntime==1.13.1
opencv-python-headless==4.6.0.66
packaging==23.0
pandas==1.5.2
Pillow==9.3.0
pooch==1.6.0
protobuf==3.20.3
pyarrow==10.0.1
pydantic==1.10.4
pydeck==0.8.0
Pygments==2.14.0
PyMatting==1.1.8
Pympler==1.0.1
pyreadline3==3.4.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
rembg==2.0.30
requests==2.28.2
rich==13.1.0
scikit-image==0.19.3
scipy==1.9.3
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.21.0
streamlit==1.17.0
sympy==1.11.1
tifffile==2022.10.10
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
uvicorn==0.20.0
validators==0.20.0
watchdog==2.1.9
yarl==1.8.2
zipp==3.11.0
